pms_recon_gui/requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
pdfplumber>=0.10.0
rapidfuzz>=3.0.0
pydantic>=2.0.0
python-dateutil>=2.8.2
pyyaml>=6.0
pytest>=7.0.0
openpyxl>=3.1.0
watchdog>=3.0.0
